Migrate AccessManagementTableRow and organize corresponding files
