+++ moderatorFrontend/src/jsMain/kotlin/webcore/MuiHelper.kt
@@ -147,3 +147,4 @@ typealias AutocompleteOnChange<T> = (SyntheticEvent<*, *>, Any, AutocompleteChan
 typealias AutocompleteOnInputChange = (event: SyntheticEvent<*, *>, value: String, reason: AutocompleteInputChangeReason) -> Unit
 typealias ButtonOnClick = (MouseEvent<HTMLButtonElement, *>) -> Unit
 typealias SelectOnChange = (ChangeEvent<HTMLInputElement>, ReactNode) -> Unit
+typealias MenuItemOnClick = () -> Unit
